fix(pie): pass PIE compilation flags to BL1 when needed
TF-A is failing to build in some situations with the error:

    ld: build/qemu/release/bl1/bl1.elf: error: PHDR segment not covered by LOAD segment

More specifically, this error happened when building TF-A v2.11
targetting the "qemu" platform, and compiling with gcc 13.3 and
binutils 2.41. This issue was observed using the Buildroot
qemu_aarch64_ebbr_defconfig in [1] with TF-A updated to v2.11. It can
be reproduced with Buildroot, with the commands:

    git clone --depth 1 --branch 2024.08-rc2 https://gitlab.com/buildroot.org/buildroot/
    cd buildroot
    make qemu_aarch64_ebbr_defconfig
    utils/config --set-str BR2_TARGET_ARM_TRUSTED_FIRMWARE_CUSTOM_VERSION_VALUE v2.11
    make olddefconfig
    make

This build failure happen because TF-A is built with PIE enabled, but
the PIE_LDFLAGS are not passed to BL1_LDFLAGS. More specifically, the
linker flag --no-dynamic-linker is missing, which will make Gnu ld run
its check and produce the described failure message.

This commit fixes the issue by setting PIE flags when needed: adding
-fpie to BL1_CPPFLAGS and BL1_CFLAGS, and adding PIE_LDFLAGS to
BL1_LDFLAGS.

Change-Id: I917dc17bf3fbae311fe3ec86a0f8089fbd2802cf
Signed-off-by: Julien Olivain <ju.o@free.fr>